services/ai-agent-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
anthropic==0.18.1
kubernetes==29.0.0
psycopg2-binary==2.9.9
requests==2.31.0
schedule==1.2.1
